requirements.txt
spotipy==2.23.0
gspread==5.12.0
google-auth==2.25.0
python-dotenv==1.0.0
black===23.9.1
flake8===6.1.0
isort==5.12.0
